+++ Tests/Functional/RenderingTest.php
@@ -7,7 +7,7 @@ use Nimut\TestingFramework\Http\Response;
 use Nimut\TestingFramework\TestCase\FunctionalTestCase;
 
 /**
- * Rendering Test Class
+ * The rendering test class.
  *
  * @package T3v\T3vContent\Tests\Functional
  */
@@ -90,11 +90,9 @@ class RenderingTest extends FunctionalTestCase {
       'ntfRoot'      => __DIR__ . '/../../.build/vendor/nimut/testing-framework/'
     ]);
 
-    $factory = \PHPUnit_Util_PHP::factory();
-
-    $response = $factory->runJob($template->render());
-
-    $result = json_decode($response['stdout'], true);
+    $php      = \PHPUnit_Util_PHP::factory();
+    $response = $php->runJob($template->render());
+    $result   = json_decode($response['stdout'], true);
 
     if ($result === null) {
       $this->fail('Frontend Response is empty:' . LF . 'Error: ' . LF . $response['stderr']);